Refactor context code
