+++ app/page.js
@@ -1,39 +1,12 @@
-'use client'
 import ProductDetails from './components/ProductDetails.jsx';
-import { AppProvider, useAppContext } from './context';
-
-const AppContent = () => {
-  const { state } = useAppContext();
-
-  if (state.loading) {
-    return <p>Loading...</p>;
-  }
-  console.log(state)
-  return (
-    <div className='antialiased'>
-      <ProductDetails />
-    </div>
-    // <div>
-    //   <h1 className='text-white'>Reviews</h1>
-    //   <ul>
-    //     {state.reviews.map((review) => {
-    //       console.log(review.text)
-    //       return (
-    //         <li key={review.id}>{review.review_body}</li>
-    //       )
-    //     })}
-    //   </ul>
-    // </div>
-  );
-};
+import Header from './components/Header.jsx';
 
 export default function Home() {
   return (
-    <div>
-      <AppProvider>
-        <AppContent />
-      </AppProvider>
-    </div>
+    <>
+      <Header />
+      <ProductDetails />
+    </>
   );
 }
